+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/forms/FormUI.kt
@@ -40,7 +40,8 @@ internal fun Form(formViewModel: FormViewModel) {
     FormInternal(
         formViewModel.hiddenIdentifiers,
         formViewModel.enabled,
-        formViewModel.elements
+        formViewModel.elements,
+        formViewModel.lastTextFieldIdentifier
     )
 }
 
@@ -48,26 +49,30 @@ internal fun Form(formViewModel: FormViewModel) {
 internal fun FormInternal(
     hiddenIdentifiersFlow: Flow<List<IdentifierSpec>>,
     enabledFlow: Flow<Boolean>,
-    elementsFlow: Flow<List<FormElement>?>
+    elementsFlow: Flow<List<FormElement>?>,
+    lastTextFieldIdentifierFlow: Flow<IdentifierSpec?>
 ) {
     val hiddenIdentifiers by hiddenIdentifiersFlow.collectAsState(emptyList())
     val enabled by enabledFlow.collectAsState(true)
     val elements by elementsFlow.collectAsState(null)
+    val lastTextFieldIdentifier by lastTextFieldIdentifierFlow.collectAsState(null)
 
     Column(
         modifier = Modifier.fillMaxWidth(1f)
 
     ) {
         elements?.let {
-            it.forEach { element ->
+            it.forEachIndexed { index, element ->
                 if (!hiddenIdentifiers.contains(element.identifier)) {
                     when (element) {
-                        is SectionElement -> SectionElementUI(enabled, element, hiddenIdentifiers)
-                        is StaticTextElement -> StaticElementUI(element)
-                        is SaveForFutureUseElement -> SaveForFutureUseElementUI(
+                        is SectionElement -> SectionElementUI(
                             enabled,
-                            element
+                            element,
+                            hiddenIdentifiers,
+                            lastTextFieldIdentifier
                         )
+                        is StaticTextElement -> StaticElementUI(element)
+                        is SaveForFutureUseElement -> SaveForFutureUseElementUI(enabled, element)
                         is AfterpayClearpayHeaderElement -> AfterpayClearpayElementUI(
                             enabled,
                             element